refactor: move delegate typedData struct to utils
